Casting change Francesco

See merge request coronavirus-outbreak-control/coronavirus-outbreak-control-android!24

+++ CoronavirusHerdImmunity/app/src/main/java/com/example/coronavirusherdimmunity/utils/BeaconDto.java
@@ -20,14 +20,14 @@ import java.util.Date;
 public class BeaconDto {
 
     public long timestmp;
-    public int identifier;
+    public Long identifier;
     public int rssi;
     public Distance distance;
     public int interval;
     public double x = 0;
     public double y = 0;
 
-    public BeaconDto(int identifier, int rssi, Distance distance, double x, double y){
+    public BeaconDto(Long identifier, int rssi, Distance distance, double x, double y){
         this.identifier = identifier;
         this.rssi = rssi;
         this.timestmp = new Date().getTime() / 1000;
@@ -36,7 +36,7 @@ public class BeaconDto {
         this.y = y;
     }
 
-    public BeaconDto(int identifier, int rssi, long timestamp, Distance distance, double x, double y){
+    public BeaconDto(Long identifier, int rssi, long timestamp, Distance distance, double x, double y){
         this.identifier = identifier;
         this.rssi = rssi;
         this.timestmp = timestamp;
@@ -46,7 +46,7 @@ public class BeaconDto {
 
     }
 
-    public BeaconDto(int identifier, int rssi, Date timestamp, Distance distance, double x, double y){
+    public BeaconDto(Long identifier, int rssi, Date timestamp, Distance distance, double x, double y){
         this.identifier = identifier;
         this.rssi = rssi;
         this.timestmp = timestamp.getTime() / 1000;
